Split the premium syncing logic into own file the PremiumSyncManager
