+++ rotkehlchen/rotkehlchen.py
@@ -1,15 +1,11 @@
 #!/usr/bin/env python
-import base64
+
 import logging
-import os
-import shutil
 import time
-from enum import Enum
-from typing import Any, Dict, List, NamedTuple, Tuple, Union
+from typing import Any, Dict, List, Tuple, Union
 
 import gevent
 from gevent.lock import Semaphore
-from typing_extensions import Literal
 
 from rotkehlchen.accounting.accountant import Accountant
 from rotkehlchen.assets.asset import Asset, EthereumToken
@@ -20,15 +16,7 @@ from rotkehlchen.blockchain import Blockchain
 from rotkehlchen.constants import SUPPORTED_EXCHANGES
 from rotkehlchen.constants.assets import A_USD, S_EUR, S_USD
 from rotkehlchen.data_handler import DataHandler
-from rotkehlchen.errors import (
-    AuthenticationError,
-    EthSyncError,
-    IncorrectApiKeyFormat,
-    InputError,
-    RemoteError,
-    RotkehlchenPermissionError,
-    UnknownAsset,
-)
+from rotkehlchen.errors import AuthenticationError, EthSyncError, InputError, UnknownAsset
 from rotkehlchen.ethchain import Ethchain
 from rotkehlchen.externalapis import Cryptocompare
 from rotkehlchen.fval import FVal
@@ -37,7 +25,8 @@ from rotkehlchen.inquirer import Inquirer
 from rotkehlchen.kraken import Kraken
 from rotkehlchen.logging import DEFAULT_ANONYMIZED_LOGS, LoggingSettings, RotkehlchenLogsAdapter
 from rotkehlchen.poloniex import Poloniex
-from rotkehlchen.premium import premium_create_and_verify
+from rotkehlchen.premium.premium import premium_create_and_verify
+from rotkehlchen.premium.sync import PremiumSyncManager
 from rotkehlchen.serializer import process_result
 from rotkehlchen.typing import (
     ApiCredentials,
@@ -55,26 +44,12 @@ from rotkehlchen.utils.misc import (
     merge_dicts,
     simple_result,
     ts_now,
-    tsToDate,
 )
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 MAIN_LOOP_SECS_DELAY = 60
-
-
-class CanSync(Enum):
-    YES = 0
-    NO = 1
-    ASK_USER = 2
-
-
-class SyncCheckResult(NamedTuple):
-    # The result of the sync check
-    can_sync: CanSync
-    # If result is ASK_USER, what should the message be?
-    message: str
 
 
 def accounts_result(per_account, totals) -> Dict:
@@ -209,84 +184,6 @@ class Rotkehlchen():
         if self.bitmex is not None:
             self.delete_exchange_data('bitmex')
 
-    def try_premium_at_start(
-            self,
-            api_key: ApiKey,
-            api_secret: ApiSecret,
-            username: str,
-            create_new: bool,
-            sync_approval: Literal['yes', 'no', 'unknown'],
-    ) -> None:
-        """Check if new user provided api pair or we already got one in the DB"""
-
-        if api_key != '':
-            assert create_new, 'We should never get here for an already existing account'
-            try:
-                self.premium = premium_create_and_verify(api_key, api_secret)
-            except (IncorrectApiKeyFormat, AuthenticationError) as e:
-                log.error('Given API key is invalid')
-                # At this point we are at a new user trying to create an account with
-                # premium API keys and we failed. But a directory was created. Remove it.
-                # But create a backup of it in case something went really wrong
-                # and the directory contained data we did not want to lose
-                shutil.move(
-                    self.user_directory,
-                    os.path.join(
-                        self.data_dir,
-                        f'auto_backup_{username}_{ts_now()}',
-                    ),
-                )
-                shutil.rmtree(self.user_directory)
-                raise AuthenticationError(
-                    'Could not verify keys for the new account. '
-                    '{}'.format(str(e)),
-                )
-
-        # else, if we got premium data in the DB initialize it and try to sync with the server
-        premium_credentials = self.data.db.get_rotkehlchen_premium()
-        if premium_credentials:
-            assert not create_new, 'We should never get here for a new account'
-            api_key = premium_credentials[0]
-            api_secret = premium_credentials[1]
-            try:
-                self.premium = premium_create_and_verify(api_key, api_secret)
-            except (IncorrectApiKeyFormat, AuthenticationError) as e:
-                log.error(
-                    f'Could not authenticate with the rotkehlchen server with '
-                    f'the API keys found in the Database. Error: {str(e)}',
-                )
-                del self.premium
-                self.premium = None
-
-        if not self.premium:
-            return
-
-        result = self.can_sync_data_from_server(new_account=create_new)
-        if result.can_sync == CanSync.ASK_USER:
-            if sync_approval == 'unknown':
-                log.info('DB data at server newer than local')
-                raise RotkehlchenPermissionError(result.message)
-            elif sync_approval == 'yes':
-                log.info('User approved data sync from server')
-                if self.sync_data_from_server_and_replace_local():
-                    if create_new:
-                        # if we successfully synced data from the server and this is
-                        # a new account, make sure the api keys are properly stored
-                        # in the DB
-                        self.data.db.set_rotkehlchen_premium(api_key, api_secret)
-            else:
-                log.debug('Could sync data from server but user refused')
-        elif result.can_sync == CanSync.YES:
-            log.info('User approved data sync from server')
-            if self.sync_data_from_server_and_replace_local():
-                if create_new:
-                    # if we successfully synced data from the server and this is
-                    # a new account, make sure the api keys are properly stored
-                    # in the DB
-                    self.data.db.set_rotkehlchen_premium(api_key, api_secret)
-
-        # else result.can_sync was no, so we do nothing
-
     def unlock_user(
             self,
             user: str,
@@ -307,13 +204,19 @@ class Rotkehlchen():
         self.password = password
         self.user_directory = self.data.unlock(user, password, create_new)
         self.last_data_upload_ts = self.data.db.get_last_data_upload_ts()
-        self.try_premium_at_start(
-            api_key=api_key,
-            api_secret=api_secret,
-            username=user,
-            create_new=create_new,
-            sync_approval=sync_approval,
-        )
+        self.premium_sync_manager = PremiumSyncManager(data=self.data, password=password)
+        try:
+            self.premium = self.premium_sync_manager.try_premium_at_start(
+                api_key=api_key,
+                api_secret=api_secret,
+                username=user,
+                create_new=create_new,
+                sync_approval=sync_approval,
+            )
+        except AuthenticationError:
+            # It means that our credentials were not accepted by the server
+            # or some other error happened
+            pass
 
         exchange_credentials = self.data.db.get_exchange_credentials()
         settings = self.data.db.get_settings()
@@ -399,136 +302,6 @@ class Rotkehlchen():
 
         self.data.set_premium_credentials(api_key, api_secret)
 
-    def maybe_upload_data_to_server(self) -> None:
-        # upload only if unlocked user has premium
-        if self.premium is None:
-            return
-
-        # upload only once per hour
-        diff = ts_now() - self.last_data_upload_ts
-        if diff > 3600:
-            self.upload_data_to_server()
-
-    def upload_data_to_server(self) -> None:
-        log.debug('upload to server -- start')
-        data, our_hash = self.data.compress_and_encrypt_db(self.password)
-        try:
-            metadata = self.premium.query_last_data_metadata()
-        except RemoteError as e:
-            log.debug(
-                'upload to server -- query last metadata failed',
-                error=str(e),
-            )
-            return
-
-        log.debug(
-            'CAN_PUSH',
-            ours=our_hash,
-            theirs=metadata.data_hash,
-        )
-        if our_hash == metadata.data_hash:
-            log.debug('upload to server -- same hash')
-            # same hash -- no need to upload anything
-            return
-
-        our_last_write_ts = self.data.db.get_last_write_ts()
-        if our_last_write_ts <= metadata.last_modify_ts:
-            # Server's DB was modified after our local DB
-            log.debug("CAN_PUSH -> 3")
-            log.debug('upload to server -- remote db more recent than local')
-            return
-
-        try:
-            self.premium.upload_data(
-                data_blob=data,
-                our_hash=our_hash,
-                last_modify_ts=our_last_write_ts,
-                compression_type='zlib',
-            )
-        except RemoteError as e:
-            log.debug('upload to server -- upload error', error=str(e))
-            return
-
-        # update the last data upload value
-        self.last_data_upload_ts = ts_now()
-        self.data.db.update_last_data_upload_ts(self.last_data_upload_ts)
-        log.debug('upload to server -- success')
-
-    def can_sync_data_from_server(self, new_account: bool) -> SyncCheckResult:
-        """
-        Checks if the remote data can be pulled from the server.
-
-        Returns a SyncCheckResult denoting whether we can pull for sure,
-        whether we can't pull or whether the user should be asked. If the user
-        should be asked a message is also returned
-        """
-        log.debug('can sync data from server -- start')
-        b64_encoded_data, our_hash = self.data.compress_and_encrypt_db(self.password)
-        try:
-            metadata = self.premium.query_last_data_metadata()
-        except RemoteError as e:
-            log.debug('can sync data from server failed', error=str(e))
-            return SyncCheckResult(can_sync=CanSync.NO, message='')
-
-        if new_account:
-            return SyncCheckResult(can_sync=CanSync.YES, message='')
-
-        if not self.data.db.get_premium_sync():
-            # If it's not a new account and the db setting for premium syncin is off stop
-            return SyncCheckResult(can_sync=CanSync.NO, message='')
-
-        log.debug(
-            'CAN_PULL',
-            ours=our_hash,
-            theirs=metadata.data_hash,
-        )
-        if our_hash == metadata.data_hash:
-            log.debug('sync from server -- same hash')
-            # same hash -- no need to get anything
-            return SyncCheckResult(can_sync=CanSync.NO, message='')
-
-        our_last_write_ts = self.data.db.get_last_write_ts()
-        if our_last_write_ts >= metadata.last_modify_ts:
-            # Local DB is newer than Server DB
-            log.debug('sync from server -- local DB more recent than remote')
-            return SyncCheckResult(can_sync=CanSync.NO, message='')
-
-        data_bytes_size = len(base64.b64decode(b64_encoded_data))
-        if data_bytes_size > metadata.data_size:
-            message_prefix = (
-                'Detected newer remote database BUT with smaller size than the local one. '
-            )
-
-        else:
-            message_prefix = 'Detected newer remote database. '
-
-        message = (
-            f'{message_prefix}'
-            f'Local size: {data_bytes_size} Remote size: {metadata.data_size} '
-            f'Local last modified time: {tsToDate(our_last_write_ts)} '
-            f'Remote last modified time: {tsToDate(metadata.last_modify_ts)} '
-            f'Would you like to replace the local DB with the remote one?'
-        )
-        return SyncCheckResult(
-            can_sync=CanSync.ASK_USER,
-            message=message,
-        )
-
-    def sync_data_from_server_and_replace_local(self) -> bool:
-        """
-        Performs syncing of data from server and replaces local db
-
-        Returns true for success and False for error/failure
-        """
-        try:
-            result = self.premium.pull_data()
-        except RemoteError as e:
-            log.debug('sync from server -- pulling failed.', error=str(e))
-            return False
-
-        self.data.decompress_and_decrypt_db(self.password, result['data'])
-        return True
-
     def start(self):
         return gevent.spawn(self.main_loop)
 
@@ -540,7 +313,7 @@ class Rotkehlchen():
             if self.kraken is not None:
                 self.kraken.main_logic()
 
-            self.maybe_upload_data_to_server()
+            self.premium_sync_manager.maybe_upload_data_to_server()
 
             log.debug('Main loop end')
